backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9
python-dotenv==1.0.1
qdrant-client==1.7.3
requests==2.31.0
edge-tts==6.1.9
fastembed==0.2.2
